Dir refactor

+++ scripts/keystore.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
+if [ -f app/keystore.gradle ]; then
+    echo "app/keystore.gradle file already exists!"
+    exit 1
+fi
 cat > app/keystore.gradle << EOF
 ext.release_keystore=file('keystore/debug.keystore')
 ext.key_alias='androiddebugkey'
